Create unfinished_Artemis_code
code is as of yet not fully operational.. I'm slightly embarased to post non happy code. mostly work needs to be done on the boron side. it records makes the files and stamps and I've got a whole i2c infrastructure going on. On the boron I've gotta get the handler to behave and do the encoding(that's easy though, i think scruff wrote encode/decode commands/libs that are compliant to whatever the particle scheme is) venkat's got fxns for that too, no need to reinvent the wheel